Resolve some issues in updating styles along with frames.
If we don't specify a style, the baseQuad is still used.  Oddly, when we
stop using the baseQuad I see some webGL warnings which I'll hunt down
separately.

It would be better to set a style of undefined if not needed (for
instance not using channel compositing).

There is one workaround of a potential caching bug that I'll also look
into later.