Reduce duplicate computation of expensive cached values.
Some item functions are cached, like histograms, tile frames, and
thumbnails.  These can be expensive.  Before, if a client asked for an
expensive function and a second client asked for the same function
before the first completed, it would compute it twice.  Now, the second
call waits for the first to complete and then reads from the cache.