Add initial DICOMweb support
This creates a new DICOMweb assetstore that is used to track resources on
a DICOMweb server. The user provides the DICOMweb server URL and optionally
QIDO and WADO prefixes (if the server requires them). The user is then able
to import references to the objects on the DICOMweb server.

The files that are created via this assetstore are able to be viewed with the
dicom tile source, which has been extended to handle files on a DICOMweb
server.

A test was also added that utilizes the public DICOMweb server located
[here](https://imagingdatacommons.github.io/slim/).

There is still more work to do in future PRs, including:

* Adding a search/filter capability
* Adding authentication for non-public servers
* Query and store metadata from the DICOMweb objects

Fixes: #1204

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>