Reduce memory allocation during some styling operations
Specifically, when we apply a style, in some instances the tile or
region is converted from one dtype to a float in order to preserve some
of the transform properties through the process.  We had been using
`float`, which internally is `np.float64`.  But, `np.float32` gives
sufficient accuracy for our styles.  With this change, `np.float32` is
used unless the base datatype is `float64`, in which we do not reduce
the precision.