Closing some styled sources could premature close file handles
This could be reproduced by opening an nd2 file, opening another
instance of it with a bad style, then trying to open another instance of
it with a good style.  The file handles would get closed with the bad
style, and then subsequent reads of a good style would fail (and even
could segfault).