Add an option to not cache sources.
When opening a tile source, pass `noCache=True`.  In this mode, the tile
source can directly have its style modified (e.g., `source.style = <new
value>`).  This is also used when importing images into girder to avoid
flushing the cache of tile sources that are in active use.

This closes #1294.

This closes #1145.

There is a config value `cache_sources`, that, if False, makes `noCache`
default to False.

This closes #1206.