Minor refactor of the how we access the tiff reader class.
There is much we are doing with libtiff that could be switch to the
python dictionaries returned by tifftools.  Would this speed things up
in common use cases?  It would make a fallback to PIL if libtiff is
unavailable easier.  Some of this refactor will make that easier to
test.