Update WsiDicomWebClient init call
As of the latest `wsidicom` release (0.13.0), the new `__init__` method
just takes a `DICOMwebClient` object. The old `__init__` method was
moved to the class method `create_client()`. So we need to update
the way we initialize it.

In #1349, we are adding authentication where we create the DICOMwebClient
object ourselves, so go ahead and do that now to simplify that one.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>