Read and set several metadata items, if available
When DICOMweb series are imported, several metadata entries from each
series are now extracted and set on the item metadata.

We are extracting most of the metadata that is displayed in the SLIM
viewer: https://imagingdatacommons.github.io/slim/

Most of the metadata entries are available as series-level metadata
items. However, some of the metadata entries are only available on
the instances themselves (in particular, the "Specimen" information,
which is nested several dataset layers deep).

The SLIM viewer sets most of its metadata by locating the first
WSI Volume dataset, and using that (it appears that the metadata
for WSI volume datasets in a series are almost completely identical
to other WSI volume datasets in the series). We are also using that
approach here.

We are using a slightly different approach for displaying the Specimen
metadata, but that is something we can always improve upon.

Additionally, the SLIM viewer is finding the first WSI Volume dataset
in a whole study, and using that for the metadata for the whole study.
We are currently doing it on a series-level, though (we find the first
WSI Volume dataset for a series, and use that to set the metadata on
the girder item that represents that series).

Fixes: #1314

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>