Remove logic for determining transfer syntax
wsidicom now determines the transfer syntax automatically, as of
version 0.14.0, so we don't need to do this anymore.

wsidicom also supports more transfer syntaxes as well, so the list
of transfer syntaxes we had are not complete now either. Just remove
this logic and let wsidicom take care of it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>