Only support Python >= 3.9 for DICOMweb
Unfortunately, the newer versions of wsidicom, which we need to support
certain data formats for DICOMweb, require Python 3.9 or greater.

So we are going to only allow the DICOMweb plugin to be loaded for
Python >= 3.9.

The regular DICOM reader will still work with wsidicom for Python 3.8,
but only the DICOMweb part will not be loaded.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>