Optimizing when reading arrays rather than images from tiff files
Specifically, if we want a numpy array, reading a jpeg or jp2k encoded
tile as the raw image bytes and then decoding it with simplejpeg or PIL
is slower than having libtiff decode it.  If, however, we want to serve
a jpeg, it is quicker to never decode it, so this code path remains the
same.