Add support for DICOMweb range requests
When a range request is made or a fuse mount is attempted, we compute the
size of the DICOM file. Currently, the only reliable way that we know to do
this is by streaming the whole file and counting the total bytes read.

After the file size is computed, it is saved in `file['size']` so it does not
need to be computed again. Then, the range requests can be successfully made,
complete with accurate `Content-Length` and `Content-Range` headers.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>