Add authentication to DICOMweb testing
This adds an additional docker-compose yaml, `auth-docker-compose.yml`, which
spins up a dcm4chee server that requires authentication via keycloak. This
authenticated server is now being used for the CircleCI testing.

This PR also adds a couple of scripts to work with this server, including an
`update_access_token_lifespan.py` script (with the default access token
lifespan of 10 minutes, the token would expire before it would be tested
in the CI), and a `create_keycloak_token.py` script, which is used to create
the access token that we provide to the assetstore.

Since we may want to perform local testing without authentication, the
non-authenticated dcm4chee docker-compose is still there. The testing code
checks for a `DICOMWEB_TEST_TOKEN` environment variable, where it expects
to find a token. If that environment variable is set, then it uses token
authentication with that token. If it is not set, then it assumes authentication
is not required and attempts to run the tests without authentication.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>