Reduce the ending check size to be ending_size - 1
If the ending is split between two chunks, then at most, the second
chunk will contain ending_size - 1 bytes of the ending. So we should
just check for this.

We also don't need to add extra bytes to check for in the ending, because
we don't care what comes after the ending.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>