Fix compositing multi source with alpha
Fix an issue with compositing sources in the multi source caused by
alpha range.  We composite tiles using pre-multiplied alpha values.
When we have 8 bit images and apply an affine transform, we do this in
float precision so that compositing does not lose precision prematurely.
The function fullAlphaValue is used to determine the range of the alpha
values we expect; for uint8 this is 255, for float it is 1.  Since the
values have been cast to float but NOT scaled, this ends up with the
wrong values.  Functionally, this was multiplying incomplete tiles by
255 and then taking the low 8 bits of the resultant integer cast, which
looked like nonsense.