Apply limit and filters to studies, not series
We were previously just performing a `search_for_series()` and applying
the limit and filters to that search.

However, it is probably more intuitive for users to be searching for
studies, rather than series. So we are now performing a `search_for_studies()`
first, applying the limit and filters to this, and then locating all
series within those studies, and proceeding from there.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>