Don't infer file size right away
Importing the first 10 studies on IDC has been taking around 20 minutes.
About 65% of that time has been spent on inferring file sizes.

Even though we don't stream the file data, the request from which we get
the content length must be taking some time on the server. Skip doing
this for now. We can add it back in if we figure out a way to make it much
faster.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>