Speed up some retiling
In one test, this reduces tiling time by about 8% when running a tile
iterator on a tile that is slightly smaller than the native size (224
square rather than 240 square).  Of the time reduction, some of the
speed is based on non-format specific gains and some specifically on the
tiff tile source; roughly 3% is the non-format specific, and 5% is tiff
specific.