Reduce bioformats keeping file handles open
When bioformats tries to open a zip file, it opens the file once per
internal file.  If it fails to find a file that can be opened within the
zip, it throws an error and leaves all of the file handles open.  By
dividing the reader into more steps, we can ask to open an empty file,
which somehow releases the file handles.