Better handle images that are composed of a folder and an item.
Some file formats (mrxs, vsi) include folders of files and a root file.
These need to be processed after all files are present, not when just
the root file is present.  This change checks for a named file within
the folder and, if present and not a large image, checks if the parent
item exists and is not a large image.