Harden the tile cache between python versions and numpy versions
If a python version with numpy >= 2.x wrote a tile to the cache, and
then a python version with numpy < 2.x tried to read the tile from the
cache, the failure was a ModuleNotFoundError rather than the errors we
were checking for.  This now counts as a cache miss rather than a
failure.