Specify empty_dir=yes when constructing vsicurl parameters
When automatically constructing a vsicurl url to allow gdal to read from
https or s3 sources, set empty_dir=yes to avoid probing for sidecar
files.  Prior to GDAL PR 11469, this seem to be the case with just
list_dir=no, but it isn't clear from the GDAL documentation why that
works.  Be explicit (open a viscurl path directly to change the
behavior).