When clearing all caches, better clear vips and gdal caches
We weren't clearing the gdal block cache, and weren't clearing the vips
cache completely.  Further, the vips operation cache can't be set to
zero or it has the potential to cause a future memory error.

GDAL also has a dataset pool (also a proxy pool), which I don't think
this clears.