Be more specific in casting when converting images via vips
If a source image was not uint8 or uint16 we used the full range of the
datatype for some of the conversion, but, to maximize compatibility, we
always output either uint8 or uint16 pixels.  In a sample image that was
stored in int32 but only had pixels with values 0, 1, 2, this resulted
it scaling the output by (val * 2 ** -16) + 2 ** 15, but in that
resulted in all output pixels having the value of 32768.

Now, we check the range of all frames of an image first, and try to
intelligently pick a datatype and scaling based on the actual range of
the data.  This will still be wrong in some instances, but should at
least result in visible images.