vips could use too much memory during image conversion
When we convert an image via vips, we collect tiled regions into strips
and then strips into a file image.  This is done because vips won't
handle more than some number of actions in its action list, but can
handle a tree of actions (each tile add being an action).  However, vips
appears to read most of the strips into memory as it goes in this memory
model.  If we save each strip as a vips temp file, we have more disk
access, but avoid the memory spike.