Improve handling of large annotation elements
We were only checking the first "few" elements to see if they were large
enough that we had to store them as files rather than in the database,
and we were not checking the size of holes in polygons.  As such, an
annotation element with a huge number of holes, or one with many small
polygons followed by a huge polygon could fail to be placed in the
database.  Further, insert errors of this sort were not being percolated
up to the logs, making debug difficult.