Add an option to merge dicom files in the same series and study
If multiple dicom files from the same series and study are uploaded or
imported to the same folder, and the option is selected, all of the
files will be moved to a single item (the first so located).  This makes
browsing WSI dicoms easier, as they appear as single items, but may have
repercussions.

This feature should be viewed as provisional until tested more
thoroughly in the field.