Show all annotations when zoomed in
When zoomed in with a power of two of the max zoom and showing
annotations, make sure all of the annotation in view are fully loaded.
Before, this was still subject to detail limits, which meant that it was
possible to fail to resolve some annotations regardless of zoom level.