Better handle images with adjacent files in the same girder item
When opening images that are uploaded or otherwise not imported from a
filesystem assetstore, we use the fuse mount to read the images.  If we
think the image is composed of adjacent files, we were using the flat
mount point, where individual items in girder each appear as single
files.  But, when the image item has multiple files that are part of the
image, this does not work.  The detection is that if the format might
have adjacent files and we are reading an unconverted file, and the item
has multiple files with the same mimetype as the main file, then we
prefer the non-flat mount.

This has the potential to fail; someone could upload multiple files and
have the same mimetype even though the image is spread across multiple
items.  But, without it, merged dicoms, for instance, may not work.