Improve reading certain tiles via tiff source
When reading certain jpeg compressed tiles and wanting them
uncompressed, we were defaulting to asking libtiff to read the
uncompressed tile in an attempt to use the more optimized path of
libtiff doing the decompressed, but sometimes this did not behave as
expected.  When so asked, if this fails, we fall back to decoding the
jpeg in our code.