Remove locking in reading nd2 tiles
When locking was added around 3 years ago, it was necessary to prevent
memory from leaking between concurrent tile reads.  Shortly after that,
we switched dask to use a single-threaded mode to avoid certain failures
to release memory (see #797, #994, #1260, and others).  Since then, both
the nd2 library and dask have improved and this is no longer necessary.
In tests with converting a large multi-frame nd2 file, is appears 4 to 5
times faster to remove these locks and thread requirements with the only
downside that more cpu is used at a time.

dask can use processes instead of threads, but this is slower.  It can
also limit the number of workers, but at the level we are running
compute, this seems to be the limit per tile and not actually useful.