Speed up annotation centroid queries
Much of the time is just getting data back from mongo.  On a ~1/3
million annotation test, before these changes it took ~4.19s to ~3.69s
(12% improvement).  If we stop sorting the results, this improves
further to ~3.50s.  Here we've stopped sorting unless it could matter.

As a further (future) improvement, we could precompute or cache the
response so the response would effectively be serving a static file.

Some experiments which were not effective:

- Switching to precomputing the data in a mongo aggregation to avoid
  some work in python.  This was slower no matter what I tried.

- Using a covering index where all values used were in the index (huge
  and not helpful)

- Changing the struct.pack in the rest response to try to use tobytes or
  numpy serialization.  This didn't change speed much and, in some
  instance, slowed things down.  It _seems_ like it would be faster to
  change how we serialized ids, but struct.pack is not the bottleneck.

- Removing the sort in the query helps.