Improve indices for annotationelement queries
Add more indices with bounding box ranges.  These are useful WHEN a
query limit is specified, which is now somewhat automatic if we specify
maxDetails as part of a query.

Don't return a count when we are using a bounding box; our current
clients do not use it and it takes some time.

Don't filter the individual fields of the bounding box, even when we
aren't using them for non-centroid queries; that filter takes a
surprising amount of time.  For centroid queries, we are reducing what
is returned about the element so reducing what is returned about the
bounding box seems to actually help.